requirements.txt
# This file lists all the dependencies that are required by the project
# Generated by pipreqs
cubo==2023.7.2
earthengine_api==0.1.371
ee==0.2
geopandas==0.14.0
joblib==1.4.2
lightgbm==3.3.3
matplotlib==3.8.0
numpy==1.23.4
pandas==2.0.3
planetary_computer==1.0.0
pyproj==3.5.0
pytorch_lightning==2.1.3
PyYAML==6.0.2
rasterio==1.3.9
rioxarray==0.15.0
scikit_learn==1.3.2
segmentation_models_pytorch==0.3.3
torch==2.3.1+cu121
torchmetrics==1.2.1
tqdm==4.66.5
wandb==0.16.2
xarray==2024.5.0
